refactor: created canMutate(Page/User) function to avoid duplicate code
